get sentence embeddings

+++ src/get_sentence_embeddings.py
@@ -1,14 +1,13 @@
 import os
-
-import plac
 import torch
-
-
+import gc
 from transformers import T5Tokenizer, T5Model
 from torch.nn import LayerNorm
 import transformers.models.t5.modeling_t5 as t5_modeling
 import numpy as np
+
 t5_modeling.T5LayerNorm = LayerNorm
+
 device = "cuda" if torch.cuda.is_available() else "cpu"
 
 from tqdm import tqdm
@@ -16,9 +15,18 @@ from tqdm import tqdm
 output_dir = "sentence_embeddings/eng_latn"
 os.makedirs(output_dir, exist_ok=True)
 
+
 def get_tokenizer_encoder(model_name="t5-base"):
+    gc.collect()
     tokenizer = T5Tokenizer.from_pretrained(model_name)
     encoder = T5Model.from_pretrained(model_name).encoder.to(device)
+
+    # Clear cache
+    torch.cuda.empty_cache()
+    if torch.cuda.is_available():
+        with torch.cuda.device(device):
+            torch.cuda.empty_cache()
+
     return tokenizer, encoder
 
 
@@ -31,9 +39,8 @@ def get_embeddings(tokenizer, encoder, data_batch, max_length=128):
 
 
 def load_data(
-data_path = "/Users/yiyichen/Documents/experiments/datasets/Morphology-Matters-corpus/eng-literal/train.txt"
-        ):
-
+        data_path="/Users/yiyichen/Documents/experiments/datasets/Morphology-Matters-corpus/eng-literal/train.txt"
+):
     print(f"Loading data from {data_path}")
     with open(data_path) as f:
         data = [x.replace("\n", "") for x in f.readlines()]
@@ -42,8 +49,7 @@ data_path = "/Users/yiyichen/Documents/experiments/datasets/Morphology-Matters-c
     return data[:5000], data[-300:]
 
 
-def get_embeddings_stack(model_name, data, max_length):
-    tokenizer, encoder = get_tokenizer_encoder(model_name)
+def get_embeddings_stack(tokenizer, encoder, data, max_length):
     embeddings_list = []
     batch_size = 50
     num_batches = int(np.ceil(len(data) / batch_size))
@@ -58,19 +64,29 @@ def get_embeddings_stack(model_name, data, max_length):
     return embeddings.detach().cpu().numpy()
 
 
-
 def main(data_path):
     train_data, test_data = load_data(data_path)
-    for model_name in ["t5-small", "t5-base", "google/flan-t5-base", "google/flan-t5-base" ]:
+    for model_name in ["t5-small", "t5-base", "google/flan-t5-base", "google/flan-t5-base"]:
+        tokenizer, encoder = get_tokenizer_encoder(model_name)
         for max_length in [32, 64, 128]:
-            print(f"processing {model_name} max length {max_length}")
-            train_embeddings = get_embeddings_stack(model_name, train_data, max_length=max_length)
-            np.save(f"{output_dir}/{model_name}_train_{max_length}.npy", train_embeddings)
+            outputfile = f"{output_dir}/{model_name}_train_{max_length}.npy"
+            if not os.path.exists(outputfile):
+                print(f"processing {model_name} max length {max_length}")
+                with torch.no_grad():
+                    train_embeddings = get_embeddings_stack(tokenizer, encoder, train_data, max_length=max_length)
+                    np.save(f"{output_dir}/{model_name}_train_{max_length}.npy", train_embeddings)
 
-            test_embeddings = get_embeddings_stack(model_name, test_data, max_length=max_length)
-            np.save(f"{output_dir}/{model_name}_test_{max_length}.npy", test_embeddings)
+                    test_embeddings = get_embeddings_stack(tokenizer, encoder, test_data, max_length=max_length)
+                    np.save(f"{output_dir}/{model_name}_test_{max_length}.npy", test_embeddings)
+
+            torch.cuda.empty_cache()
+        del encoder
+        del tokenizer
+        gc.collect()
+        torch.cuda.empty_cache()
 
 
 if __name__ == '__main__':
     import plac
+
     plac.call(main)
